Moved section utils.
