+++ src/engine/core/ui/debug/sections/DebugItemsSection.ts
@@ -12,7 +12,8 @@ import {
   getWeaponSections,
 } from "@/engine/core/utils/ini";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { getInventoryNameForItemSection, isAmmoSection, spawnItemsForObject } from "@/engine/core/utils/object";
+import { getInventoryNameForItemSection, spawnItemsForObject } from "@/engine/core/utils/object";
+import { isAmmoSection } from "@/engine/core/utils/section";
 import { resolveXmlFile } from "@/engine/core/utils/ui";
 import { TInventoryItem } from "@/engine/lib/constants/items";
 import { drugs } from "@/engine/lib/constants/items/drugs";
