+++ src/engine/core/utils/camp.test.ts
@@ -1,0 +1,13 @@
+import { describe, it } from "@jest/globals";
+
+describe("camp utils", () => {
+  it.todo("startPlayingGuitar should correctly start playing");
+
+  it.todo("startPlayingHarmonica should correctly start playing");
+
+  it.todo("canTellCampStory should correctly check");
+
+  it.todo("canPlayCampGuitar should correctly check");
+
+  it.todo("canPlayCampHarmonica should correctly check");
+});
