+++ src/engine/core/utils/object/index.ts
@@ -6,7 +6,6 @@ export * from "@/engine/core/utils/object/object_get";
 export * from "@/engine/core/utils/object/object_location";
 export * from "@/engine/core/utils/object/object_loot";
 export * from "@/engine/core/utils/object/object_meeting";
-export * from "@/engine/core/utils/object/object_section";
 export * from "@/engine/core/utils/object/object_set";
 export * from "@/engine/core/utils/object/object_setup";
 export * from "@/engine/core/utils/object/object_smart_cover";
